Add Generic Repository Pattern with Specification Support
Introduced a `GenericRepository` class to implement a generic
repository pattern for data access, supporting CRUD operations
and leveraging Entity Framework Core. Added logging for
traceability.

Defined an `IDbEntity` interface to enforce a primary key
property on entities. Created an `IGenericRepository` interface
to specify the contract for the repository.

Added an `IIncludeSpecification` interface and an abstract
`IncludeSpecification` class to support the specification
pattern for including related entities in queries.